[CREATE] GroupManagement page dev completed
